Replaced mocked types with instantiations of concrete type => resulted in failed tests
