+++ Specifications/Applications/for_ApplicationArtifactIdentifierStringConverter/for_application_with_bounded_context_module_feature_and_recursive_sub_feature/when_converting_identifier_with_five_segments_for_a_resource_to_string.cs
@@ -7,7 +7,7 @@ using Machine.Specifications;
 using Moq;
 using It = Machine.Specifications.It;
 
-namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter
+namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_with_bounded_context_module_feature_and_recursive_sub_feature
 {
     public class when_converting_identifier_with_five_segments_for_a_resource_to_string : given.an_application_resource_identifier_converter
     {
