+++ Specifications/Applications/for_ApplicationArtifactIdentifierStringConverter/for_application_with_bounded_context_module_feature_and_recursive_sub_feature/when_converting_string_identifier_missing_artifact_generation.cs
@@ -1,7 +1,7 @@
 using System;
 using Machine.Specifications;
 
-namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter
+namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_with_bounded_context_module_feature_and_recursive_sub_feature
 {
     public class when_converting_string_identifier_missing_artifact_generation : given.an_application_resource_identifier_converter
     {
