+++ Specifications/Applications/for_ApplicationArtifactIdentifierStringConverter/for_application_with_bounded_context_module_feature_and_recursive_sub_feature/when_converting_string_identifier_with_bounded_context_module_feature_and_sub_feature.cs
@@ -6,21 +6,23 @@ using System.Linq;
 using Machine.Specifications;
 using It = Machine.Specifications.It;
 
-namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter
+namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringConverter.for_application_with_bounded_context_module_feature_and_recursive_sub_feature
 {
-    public class when_converting_string_identifier_with_only_bounded_context : given.an_application_resource_identifier_converter
+    public class when_converting_string_identifier_with_bounded_context_module_feature_and_sub_feature : given.an_application_resource_identifier_converter
     {
         const string bounded_context_name = "TheBoundedContext";
         const string module_name = "TheModule";
         const string feature_name = "TheFeature";
         const string sub_feature_name = "TheSubFeature";
-        const string second_level_sub_feature_name = "TheSecondLevelSubFeature";
         const string resource_name = "MyResource";
         const int artifact_generation = 1;
 
         static string string_identifier =
             $"{application_name}{ApplicationArtifactIdentifierStringConverter.ApplicationSeparator}" +
             $"{bounded_context_name}{ApplicationArtifactIdentifierStringConverter.ApplicationLocationSeparator}" +
+            $"{module_name}{ApplicationArtifactIdentifierStringConverter.ApplicationLocationSeparator}" +
+            $"{feature_name}{ApplicationArtifactIdentifierStringConverter.ApplicationLocationSeparator}" +
+            $"{sub_feature_name}" +
             $"{ApplicationArtifactIdentifierStringConverter.ApplicationArtifactSeparator}{resource_name}"+
             $"{ApplicationArtifactIdentifierStringConverter.ApplicationArtifactTypeSeparator}{artifact_type_name}"+
             $"{ApplicationArtifactIdentifierStringConverter.ApplicationArtifactGenerationSeperator}{artifact_generation}";
@@ -30,9 +32,13 @@ namespace Dolittle.Applications.Specs.for_ApplicationArtifactIdentifierStringCon
         Because of = () => identifier = converter.FromString(string_identifier);
 
         It should_return_a_matching_identifier = () => identifier.ShouldNotBeNull();
-        It should_hold_the_application = () => identifier.Application.ShouldEqual(application.Object);
+        It should_hold_the_application = () => identifier.Application.ShouldEqual(application);
         It should_hold_the_artifact = () => identifier.Artifact.Name.Value.ShouldEqual(resource_name);
-        It should_hold_the_bounded_context_segment = () => identifier.Location.Segments.First().Name.AsString().ShouldEqual(bounded_context_name);
+        It should_hold_the_four_segments = () => identifier.Location.Segments.Count().ShouldEqual(4);
+        It should_hold_the_bounded_context_segment = () => identifier.Location.Segments.ToArray()[0].Name.AsString().ShouldEqual(bounded_context_name);
+        It should_hold_the_module_segment = () => identifier.Location.Segments.ToArray()[1].Name.AsString().ShouldEqual(module_name);
+        It should_hold_the_feature_segment = () => identifier.Location.Segments.ToArray()[2].Name.AsString().ShouldEqual(feature_name);
+        It should_hold_the_sub_feature_segment = () => identifier.Location.Segments.ToArray()[3].Name.AsString().ShouldEqual(sub_feature_name);
         It should_hold_the_artifact_type = () => identifier.Artifact.Type.ShouldEqual(artifact_type.Object);
         It should_hold_the_artifact_generation = () => identifier.Artifact.Generation.Value.ShouldEqual(artifact_generation);
     }
